+++ docs/README.md
@@ -3,16 +3,20 @@
 ## 1. 로또 구매하기
 
 - [x] 로또 구입 금액을 입력 받는다. ( 1,000원 단위 )
+
   - [x] 숫자가 아닌 경우
   - [x] 양의 정수가 아닌 경우
   - [x] 1000 단위로 나누어 떨어지지 않는 경우
   - [x] 한 장의 로또도 구매할 수 없는 경우 ( 1,000원 미만 )
   - [x] 최대 입력값 제한 : 10만원
-- [ ] 해당 금액으로 살 수 있는 로또의 개수를 계산한다. ( 입력값 / 1000 )
-- [ ] 랜덤 로또 번호를 지닌 로또를 생성하는 기능을 만든다.
-  - **중복되지 않는 6개의 숫자**를 뽑는다
-  - 로또 번호의 **숫자 범위는 1~45**까지이다.
-- [ ] 살 수 있는 만큼 로또를 구매(발행)한다.
+
+- [x] 로또를 생성한다.
+
+  - [x] 1 ~ 45 사이 숫자가 아닌 경우
+  - [x] 중복되는 경우
+  - [x] 번호가 6개가 아닌 경우
+
+- [ ] 살 수 있는 만큼 ( 구입 금액 / 1000 ) 로또를 구매(발행)한다.
 - [ ] 발행한 로또 수량 및 번호를 출력한다.
   ```
   8개를 구매했습니다.
